requirements.txt
black==24.8.0
bs4==0.0.2
click==8.1.7
geographiclib==2.0
geopy==2.4.1
idna==3.8
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.3.3
PyQt5==5.15.10
python-dotenv==1.0.1
setuptools==72.1.0
urllib3==2.2.3
wheel==0.44.0
httpx==0.27.0